+++ project_2.sdk/project2/src/audio_fifo.h
@@ -8,8 +8,10 @@
 #ifndef SRC_AUDIO_FIFO_H_
 #define SRC_AUDIO_FIFO_H_
 
+#include <stdint.h>
+
 #define FSL_ID 0 //It's the only FSL so Xilinx didn't generate an ID in xparameters.h
 
-void push_to_fifo(uint16_t data);
+void push_to_fifo(uint32_t data);
 
 #endif /* SRC_AUDIO_FIFO_H_ */
